# src/geek-microservices-2021/MoviesClientApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CatalogApi",
    "version": "v1"
  },
  "paths": {
    "/api/Movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "GetAllMovies",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movies"
        ],
        "operationId": "AddNewMovie",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
    "/api/Movies/{id}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "GetMovieById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Movies"
        ],
        "operationId": "DeleteMovie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}